Create a BreakpointItem component to make Breakpoints component cleaner
